+++ coreactivity.php
@@ -7,7 +7,7 @@
  * Author URI:        https://www.dev4press.com/
  * Text Domain:       coreactivity
  * Version:           1.6
- * Requires at least: 5.7
+ * Requires at least: 5.8
  * Tested up to:      6.4
  * Requires PHP:      7.4
  * Network:           true